fix(export): mixed-precision — vision/projector fp16, LLM bf16
77c203f changed ALL export dummies to bf16. The LLM is precision-sensitive and
needs bf16. The vision and projector engines are NOT (cos_sim 0.989/0.996
against PT) and we want to keep them fp16 for build-time + numpy I/O reasons.
Forcing them to bf16 broke runtime:

  File ".../trt/engine.py", line 136, in _coerce_input
  TypeError: BF16 binding 'pixel_values': pass torch.bfloat16 tensor or
  np.uint16 buffer; got numpy float16

Because the orchestrator passes pixel_values as np.float16 to the vit_engine,
but engine.py refuses fp16 numpy when the binding is bf16 (silent
fp16→bf16 byte reinterpretation would corrupt the input — fp16's 1-5-10
layout is incompatible with bf16's 1-8-7).

Fix:

  export/vision.py, export/projector.py:
    dtype default torch.bfloat16 -> torch.float16  (REVERTED to fp16; these
    engines stay fp16-bound, no behaviour change vs pre-77c203f)

  export/llm.py:
    Stays bf16 (export_llm_prefill, export_llm_decode, export_llm_decode_ar)

  orchestrator.export_engines:
    Cast ONLY vision_model + mlp1 to fp16 around their respective exports
    (sandwich pattern: cast → export → restore). language_model stays at
    REF_DTYPE (bf16) throughout. PT inference is unaffected since the
    submodule cast is reverted before the function returns.

    The finally block now defensively restores both submodules in case an
    export raised mid-stream. Removes the global model.to(fp16) cast that
    77c203f had eliminated — we never want the LLM in fp16, only the small
    vision/projector subgraphs during their export traces.

The fp16 (visual_features from projector) → bf16 (prefill input binding)
boundary at _TRTGenerator._prefill_trt was already handled in 77c203f via
_np16_to_bf16_torch (real precision cast through fp32). That logic stays.

Tests: 113 passed.

User: pip reinstall + restart Colab + Run All. Fingerprint v3 already in
place — engines re-export with the corrected mixed-precision contract.
Vision/projector engines unchanged from the working v2 build; LLM engines
get bf16 I/O for the first time (where it actually matters).

+++ lrai_locate_anything/export/vision.py
@@ -64,7 +64,7 @@ def export_vision(
     grid_h: int,
     grid_w: int,
     onnx_path: Path,
-    dtype: torch.dtype = torch.bfloat16,
+    dtype: torch.dtype = torch.float16,
     device: str = "cuda",
 ) -> Path:
     """Export VisionForExport(vit_model, grid_h, grid_w) to ONNX at onnx_path.
